refactor: Move `asBlockId` function to utils
